(ctor, grow_heap): fixed the fix by reverting the types to size_t and casting (i + 1) to int before negating

+++ trunk/ace/Timer_Heap.cpp
@@ -1,3 +1,5 @@
+// $Id$
+
 #define ACE_BUILD_DLL
 
 #include "ace/Timer_Heap.h"
@@ -46,8 +48,8 @@ ACE_Timer_Heap::ACE_Timer_Heap (size_t size,
   // Initialize the "freelist," which uses negative values to
   // distinguish freelist elements from "pointers" into the <heap_>
   // array.
-  for (int i = 0; i < size; i++)
-    this->timer_ids_[i] = -(i + 1);
+  for (size_t i = 0; i < size; i++)
+    this->timer_ids_[i] = -((int) (i + 1));
 
   if (preallocate)
     {
@@ -294,8 +296,8 @@ ACE_Timer_Heap::grow_heap (void)
    this->timer_ids_ = new_timer_ids;
 
    // and add the new elements to the end of the "freelist"
-   for (int i = this->max_size_; i < new_size; i++)
-     this->timer_ids_[i] = -(i + 1);
+   for (size_t i = this->max_size_; i < new_size; i++)
+     this->timer_ids_[i] = -((int) (i + 1));
 
    // Grow the preallocation array (if using preallocation)
    if (this->preallocated_nodes_ != 0)
